requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
PyMySQL==1.0.2
<<<<<<< HEAD
python-dotenv==0.19.0cryptography==44.0.2
=======
python-dotenv==0.19.0
>>>>>>> miracle/main
